requirements.txt
python-ldap==3.2.0
psycopg2==2.8.4
pyyaml==5.2